requirements.txt
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
fastapi==0.95.2
redis==4.5.4
kafka-python==2.0.2
pytest==7.3.1
mlflow==2.3.1
pyyaml==6.0
lightgbm==3.3.1